requirements.txt
streamlit==1.25.0
pandas==2.3.3
mysql-connector-python==9.5.0
plotly==5.16.1
pillow==9.5.0
SQLAlchemy==2.0.20
python-dotenv==1.0.1
